fix: hashmap.loadFactor as float instead of int resize is working back ...

+++ hashmap/include/hashmap.h
@@ -24,7 +24,7 @@ typedef struct {
 
 typedef struct {
     int capacity;
-    int loadFactor;
+    float loadFactor;
     int keyCount;
 
     Node_t **table;
